move preconfigured_department to conftest
