+++ tests/conftest.py
@@ -6,6 +6,8 @@ from webtest import TestApp
 from comport.settings import TestConfig
 from comport.app import create_app
 from comport.database import db as _db
+from comport.content.models import ChartBlock
+from comport.department.models import Department
 
 from .factories import UserFactory, DepartmentFactory
 
@@ -43,3 +45,69 @@ def user(db):
     user.save()
     db.session.commit()
     return user
+
+@pytest.fixture
+def preconfigured_department():
+    # create a department
+    department = Department.create(name="Good Police Department", short_name="GPD", load_defaults=False)
+
+    # create & append assaults chart blocks with the expected slugs
+    assaults_intro = ChartBlock(title="INTRO", dataset="intros", slug="assaults-introduction", content="AAAAAAAAAAAAAA")
+    assaults_bst = ChartBlock(title="BYSERVICETYPE", dataset="byservicetype", slug="assaults-by-service-type", content="AAAAAAAAAAAAAA")
+    assaults_bft = ChartBlock(title="BYFORCETYPE", dataset="byforcetype", slug="assaults-by-force-type", content="AAAAAAAAAAAAAA")
+    assaults_bof = ChartBlock(title="BYOFFICER", dataset="byofficer", slug="assaults-by-officer", content="AAAAAAAAAAAAAA")
+
+    department.chart_blocks.append(assaults_intro)
+    department.chart_blocks.append(assaults_bst)
+    department.chart_blocks.append(assaults_bft)
+    department.chart_blocks.append(assaults_bof)
+
+    # create & append complaint chart blocks with the expected slugs
+    complaint_intro = ChartBlock(title="INTRO", dataset="intros", slug="complaints-introduction", content="BBBBBBBBBBBBB")
+    complaint_bm = ChartBlock(title="BYMONTH", dataset="bymonth", slug="complaints-by-month", content="BBBBBBBBBBBBB")
+    complaint_bya = ChartBlock(title="BYALLEGATION", dataset="bya", slug="complaints-by-allegation", content="BBBBBBBBBBBBB")
+    complaint_byat = ChartBlock(title="BYALLEGATIONTYPE", dataset="byat", slug="complaints-by-allegation-type", content="BBBBBBBBBBBBB")
+    complaint_bdis = ChartBlock(title="BYDISPOSITION", dataset="bdis", slug="complaints-by-disposition", content="BBBBBBBBBBBBB")
+    complaint_bpre = ChartBlock(title="BYPRECINCT", dataset="bpre", slug="complaints-by-precinct", content="BBBBBBBBBBBBB")
+    complaint_od = ChartBlock(title="OFFICERDEMOS", dataset="od", slug="officer-demographics", content="BBBBBBBBBBBBB")
+    complaint_bde = ChartBlock(title="BYDEMO", dataset="bde", slug="complaints-by-demographic", content="BBBBBBBBBBBBB")
+    complaint_bof = ChartBlock(title="BYOFFICER", dataset="bof", slug="complaints-by-officer", content="BBBBBBBBBBBBB")
+
+    department.chart_blocks.append(complaint_intro)
+    department.chart_blocks.append(complaint_bm)
+    department.chart_blocks.append(complaint_bya)
+    department.chart_blocks.append(complaint_byat)
+    department.chart_blocks.append(complaint_bdis)
+    department.chart_blocks.append(complaint_bpre)
+    department.chart_blocks.append(complaint_od)
+    department.chart_blocks.append(complaint_bde)
+    department.chart_blocks.append(complaint_bof)
+
+    # create & append use of force chart blocks with the expected slugs
+    uof_intro = ChartBlock(title="INTRO", dataset="intros", slug="uof-introduction", content="CCCCCCCCCCCCCC")
+    uof_ft = ChartBlock(title="FORCETYPE", dataset="forcetype", slug="uof-force-type", content="CCCCCCCCCCCCCC")
+    uof_bid = ChartBlock(title="BYINCDISTRICT", dataset="bid", slug="uof-by-inc-district", content="CCCCCCCCCCCCCC")
+    uof_od = ChartBlock(title="OFFICERDEMOS", dataset="od", slug="officer-demographics", content="CCCCCCCCCCCCCC")
+    uof_race = ChartBlock(title="RACE", dataset="race", slug="uof-race", content="CCCCCCCCCCCCCC")
+
+    department.chart_blocks.append(uof_intro)
+    department.chart_blocks.append(uof_ft)
+    department.chart_blocks.append(uof_bid)
+    department.chart_blocks.append(uof_od)
+    department.chart_blocks.append(uof_race)
+
+    # create & append officer involved shooting chart blocks with the expected slugs
+    ois_intro = ChartBlock(title="INTRO", dataset="intros", slug="ois-introduction", content="DDDDDDDDDDDDDDD")
+    ois_bid = ChartBlock(title="BYINCDISTRICT", dataset="bid", slug="ois-by-inc-district", content="DDDDDDDDDDDDDDD")
+    ois_wt = ChartBlock(title="WEAPONTYPE", dataset="weapontype", slug="ois-weapon-type", content="DDDDDDDDDDDDDDD")
+    ois_od = ChartBlock(title="OFFICERDEMOS", dataset="od", slug="officer-demographics", content="DDDDDDDDDDDDDDD")
+    ois_race = ChartBlock(title="RACE", dataset="race", slug="ois-race", content="DDDDDDDDDDDDDDD")
+
+    department.chart_blocks.append(ois_intro)
+    department.chart_blocks.append(ois_bid)
+    department.chart_blocks.append(ois_wt)
+    department.chart_blocks.append(ois_od)
+    department.chart_blocks.append(ois_race)
+
+    department.save()
+    return department, assaults_intro
